[Doc] update demo
