--- docs/src/demo-entry.js
@@ -1,63 +1,0 @@
-/* eslint-disable */
-// This file is auto gererated by build/build-entry.js
-import { wrapper } from './demo-common';
-
-export default {
-  'action-sheet': () => wrapper(import('../../packages/action-sheet/demo'), 'action-sheet'),
-  'address-edit': () => wrapper(import('../../packages/address-edit/demo'), 'address-edit'),
-  'address-list': () => wrapper(import('../../packages/address-list/demo'), 'address-list'),
-  'area': () => wrapper(import('../../packages/area/demo'), 'area'),
-  'button': () => wrapper(import('../../packages/button/demo'), 'button'),
-  'card': () => wrapper(import('../../packages/card/demo'), 'card'),
-  'cell': () => wrapper(import('../../packages/cell/demo'), 'cell'),
-  'checkbox': () => wrapper(import('../../packages/checkbox/demo'), 'checkbox'),
-  'circle': () => wrapper(import('../../packages/circle/demo'), 'circle'),
-  'col': () => wrapper(import('../../packages/col/demo'), 'col'),
-  'collapse': () => wrapper(import('../../packages/collapse/demo'), 'collapse'),
-  'contact-card': () => wrapper(import('../../packages/contact-card/demo'), 'contact-card'),
-  'coupon-list': () => wrapper(import('../../packages/coupon-list/demo'), 'coupon-list'),
-  'datetime-picker': () => wrapper(import('../../packages/datetime-picker/demo'), 'datetime-picker'),
-  'dialog': () => wrapper(import('../../packages/dialog/demo'), 'dialog'),
-  'dropdown-menu': () => wrapper(import('../../packages/dropdown-menu/demo'), 'dropdown-menu'),
-  'field': () => wrapper(import('../../packages/field/demo'), 'field'),
-  'goods-action': () => wrapper(import('../../packages/goods-action/demo'), 'goods-action'),
-  'icon': () => wrapper(import('../../packages/icon/demo'), 'icon'),
-  'image': () => wrapper(import('../../packages/image/demo'), 'image'),
-  'image-preview': () => wrapper(import('../../packages/image-preview/demo'), 'image-preview'),
-  'index-bar': () => wrapper(import('../../packages/index-bar/demo'), 'index-bar'),
-  'lazyload': () => wrapper(import('../../packages/lazyload/demo'), 'lazyload'),
-  'list': () => wrapper(import('../../packages/list/demo'), 'list'),
-  'loading': () => wrapper(import('../../packages/loading/demo'), 'loading'),
-  'nav-bar': () => wrapper(import('../../packages/nav-bar/demo'), 'nav-bar'),
-  'notice-bar': () => wrapper(import('../../packages/notice-bar/demo'), 'notice-bar'),
-  'notify': () => wrapper(import('../../packages/notify/demo'), 'notify'),
-  'number-keyboard': () => wrapper(import('../../packages/number-keyboard/demo'), 'number-keyboard'),
-  'pagination': () => wrapper(import('../../packages/pagination/demo'), 'pagination'),
-  'panel': () => wrapper(import('../../packages/panel/demo'), 'panel'),
-  'password-input': () => wrapper(import('../../packages/password-input/demo'), 'password-input'),
-  'picker': () => wrapper(import('../../packages/picker/demo'), 'picker'),
-  'popup': () => wrapper(import('../../packages/popup/demo'), 'popup'),
-  'progress': () => wrapper(import('../../packages/progress/demo'), 'progress'),
-  'pull-refresh': () => wrapper(import('../../packages/pull-refresh/demo'), 'pull-refresh'),
-  'radio': () => wrapper(import('../../packages/radio/demo'), 'radio'),
-  'rate': () => wrapper(import('../../packages/rate/demo'), 'rate'),
-  'search': () => wrapper(import('../../packages/search/demo'), 'search'),
-  'sidebar': () => wrapper(import('../../packages/sidebar/demo'), 'sidebar'),
-  'skeleton': () => wrapper(import('../../packages/skeleton/demo'), 'skeleton'),
-  'sku': () => wrapper(import('../../packages/sku/demo'), 'sku'),
-  'slider': () => wrapper(import('../../packages/slider/demo'), 'slider'),
-  'stepper': () => wrapper(import('../../packages/stepper/demo'), 'stepper'),
-  'steps': () => wrapper(import('../../packages/steps/demo'), 'steps'),
-  'style': () => wrapper(import('../../packages/style/demo'), 'style'),
-  'submit-bar': () => wrapper(import('../../packages/submit-bar/demo'), 'submit-bar'),
-  'swipe': () => wrapper(import('../../packages/swipe/demo'), 'swipe'),
-  'swipe-cell': () => wrapper(import('../../packages/swipe-cell/demo'), 'swipe-cell'),
-  'switch': () => wrapper(import('../../packages/switch/demo'), 'switch'),
-  'switch-cell': () => wrapper(import('../../packages/switch-cell/demo'), 'switch-cell'),
-  'tab': () => wrapper(import('../../packages/tab/demo'), 'tab'),
-  'tabbar': () => wrapper(import('../../packages/tabbar/demo'), 'tabbar'),
-  'tag': () => wrapper(import('../../packages/tag/demo'), 'tag'),
-  'toast': () => wrapper(import('../../packages/toast/demo'), 'toast'),
-  'tree-select': () => wrapper(import('../../packages/tree-select/demo'), 'tree-select'),
-  'uploader': () => wrapper(import('../../packages/uploader/demo'), 'uploader')
-};
