+++ docs/src/utils/index.js
@@ -20,7 +20,14 @@ function iframeReady(iframe, callback) {
 const ua = navigator.userAgent.toLowerCase();
 const isMobile = /ios|iphone|ipod|ipad|android/.test(ua);
 
+function importAll(map, r) {
+  r.keys().forEach(key => {
+    map[key] = r(key);
+  });
+}
+
 export {
   isMobile,
+  importAll,
   iframeReady
 };
